requirements.txt
colorcet==3.0.1
hydra-core==1.3.2
jupyter==1.0.0
matplotlib==3.5.2
mlflow==1.25.1
networkx==2.8
numpy==1.22.3
notebook==6.4.12
omegaconf==2.2.3
pandas==1.4.2
PhenoGraph==1.5.7
protobuf==3.20
pynvml==11.4.1
pytest==7.1.0
pytorch_lightning==1.6.4
PyYAML==6.0
pyarrow==9.0.0
ray[tune]==2.0.0
scikit_learn==1.1.2
scikit-image==0.19.3
scipy==1.8.0
seaborn==0.12.0
torch==1.11.0
torcheval==0.0.7
tqdm==4.64.0
umap==0.1.1
umap_learn==0.5.3
black[jupyter]==22.6.0
isort==5.10
pipreqs==0.4.11
pylint==2.14.5
flake8==4.0.1
pre-commit==2.9.2
pytest-cov==3.0.0
pytest-mock==3.10.0
mypy==1.0
einops==0.7.0
Pillow==10.0.1
